Back out subunit2junitxml exit status change from ticket:3336
I don't know why I included this change.  It was certainly based on a faulty
assumption (that the tox command would exit with error if a test failed - it
won't).  Apart from that, the motivation was the Python 3.6 ratchet job.  The
test run was expected to fail but then the *ratchet* tool was in charge of
deciding if this indicated real failure or not.  Failing at this point
prevented that logic from running.  Now the ratchet is gone and we want CI to
turn red if anything fails.