Replace eliot_logged_test with something ... else
The previous implementation relied on TestCase.addCleanup in an unreliable
way.

This implementation pushes the decoration logic in to the test method itself
on the premise that test methods that do weird stuff are better supported than
`run` methods that do weird stuff.

Also add some more tests for this behavior.