Use latest coveralls-python available
We've been using coveralls-1.11.1 with Python 2.7, it seems, and
coveralls-3.0.1 with Python 3.x.  Perhaps it is the older version of
coveralls that's been "pending completion" statuses such as the ones
in below?

https://coveralls.io/builds/37720929

Note that the corresponding jobs (although with different build# for
some reason) for Python 3.6 have received completion status:

https://coveralls.io/builds/37721173