Quote some YAML strings
Following PR review feedback.

Some parts of GitHub Actions configuration follows convention in GA's
documentation, in which YAML strings are not quoted, but that probably
is not a good idea.  We also don't want to change all the strings in
this unrelated set of changes.  So we will quote strings as we go, in
the blocks we touch.