GitHub Actions: really use x86 Python on Windows
In a prior commit, I mistakenly used "architecture: x64" instead of
"architecture: x86", and tests actually passed.

That was surprising, because netifaces do not have amd64_win will on
PyPI.  But mystery was solved when itamarst pointed out that netifaces
wheel (that we previously built) is present on pip cache.

But pip cache might be purged one day, and tests will fail again that
day.  We can't have that, so we will try to stick with x86 for now.