remove the extra mutable-only behavior
There seems to be no reason that this should be mutable-only.  However, the
current immutable lease management APIs do not expose the necessary
information.  Also, it seems highly unlikely that any clients anywhere are
making use of the information provided in the mutable case.

Therefore, punt on this for now.  A future version of GBS may re-introduce it
for both mutables and immutables.