Get rid of `check_memory` which depends on the control port
This was some kind of memory usage analysis tool.  It depends on the control
port so it cannot work after I delete the control port.  The code itself is
messy, undocumented, and has no automated tests.  I don't know if it works at
all anymore.  Even if it does, no one ever runs it.

Measuring Tahoe-LAFS' memory usage over the course of maintenance and
development is a lovely idea but the project has not managed to adopt (or
maintain?) that practice based on this tool.  Given sufficient interest we can
resurrect this idea using a more streamlined process and less invasive tools
in the future.