Add a check
Just one version of the check because there seems to be no way to select which
checks `nix flake check` runs and running the test suite 4 times for that
seems like a bummer.

Also removing python310 which fails to build because `poetry` doesn't get put
in the right build environment of some dependency for some reason there.