Retry running coverage tests
Sometimes these tests are moody, and re-running them seems to work.
Sometimes these tests seems to "spin" for a long time, and in those
cases it is probably better to let them fail earlier without a timeout
and retry.

A timeout of 40 minutes seem to be reasonable, based on a quick look
at the time taken to run `python -m tox` step:

- windows-latest, 3.7: 23m 12s
- windows-latest, 3.8: 29m 35s
- windows-latest, 3.9: 26m 5s
- windows-latest, 3.10: 25m 32s

- ubuntu-latest, 3.7: 20m 40s
- ubuntu-latest, 3.8: 19m 7s
- ubuntu-latest, 3.9: 18m 3s
- ubuntu-latest, 3.10: 19m 19s

- macos-latest, 3.9: 32m 7s
- macos-latest, 3.10: 38m 41s

- ubuntu-latest, pypy-3.7: 24m 47s
- ubuntu-latest, pypy-3.8: 19m 13s

(from https://github.com/tahoe-lafs/tahoe-lafs/actions/runs/3533138743)