Try a different tag
https://circleci.com/developer/machine/image/windows-server-2022-gui says
"2023.06.1" is a valid tag but real execution says "Job was rejected because
resource class medium, image windows-server-2022-gui:2023.06.1 is not a valid
resource class".  Is it even complaining about the image tag?  Or is it
complaining about the resource class?  I don't know.  I didn't touch the
resource class though.