sych with Upstream (#9)
* YOLO windows image

* Move tool versions step down a bit to where we actually have all these tools

* Work around Windows 2022 _version.py writing race

* Flick a byte in setup.py in order to rotate the CircleCI cache key

* We need the _version.py file in any case, even if we use the cache.

* Windows server 2022 current image has python 3.9 and 3.12 installed

* cleanup

* Add news.

* grammar

* This pacifies mypy regarding treq/StubTreq for me

Lots of trial and error, incl. trying the latest treq which
ships type annotations... in the end I just removed the stuff
mypy was complaining about and it made the errors go away.
Can it be as simple as that?

* MyPy in tests: Ignore StubTreq being a stub and not having all the attrs

* Remove now-unused imports to fix codechecks

* Add news

* Integration tests: 5min instead of 3min timeout

Recently, mostly on CircleCI, the CI test suite runs into
timeouts, often leading to subsequent "Dirty Reactor" errors.

Increasing the timeout makes this happen less often.

Refs ticket: 4085, 4022, 2994

* News

* Revert "MyPy in tests: Ignore StubTreq being a stub and not having all the attrs"

This reverts commit 7efc31d55246fdf57968862c6f36fa34c2a14585.

Greetings @meejah!  This is mostly to show you (in CI) the type error that MyPy
reports for StubTreq in the tests and that it has nothing to do with the
earlier changing changing type annotations in the code.

* Different fix for MyPy complaining about StubTreq

* Revert "Different fix for MyPy complaining about StubTreq"

This reverts commit 1985eaac01a010fd27af6640d42e86b46d587fa2.

It was pacifying MyPy but breaking the unit test at runtime.

* MyPy in tests: Ignore StubTreq being a stub and not having all the attrs

* remove need for mypy override

* GitHub Actions: Don't run the Pypy 3.8 job anymore.

Fixes ticket: 4130.

* Add (no) news.

* upload-artifact -> v4

* hopefully unique names (upload v4)

* Add (no) news.

---------

Co-authored-by: meejah <meejah@meejah.ca>
Co-authored-by: Florian Sesser <florian@leastauthority.com>