Upload test report artifact
As a temporary debugging step for now, but we may need to keep it so
that the next step will work for forked PRs also.  This is because
forked PRs run with fewer permissions, so test-reporter@v2 can't write
its results to our repository.  So we will upload the XML, and then
analyze it in a separate workflow with the required permissions.